requirements.txt
requests==2.18.4
flask==1.0.2
redis==2.10.6
PyYAML==3.12
gunicorn==19.8.1
pytest==3.6.0
daiquiri==1.3.0
selenium==3.12.0
